+++ cc2dataset/main.py
@@ -17,199 +17,11 @@ import math
 import time
 from .spark_session_builder import build_spark_session
 from io import BytesIO
+from .wat_utils import process_wat
+from .warc_utils import process_warc
+from .index_utils import read_index_files
 
 
-def valid_video_link(link):
-    valid_http = link.get("url", "").startswith("http")
-    valid_video = any(
-        link.get("url", "").endswith(ext) for ext in [".avi", ".mp4", ".mkv", ".webm", ".mov", ".mpg", ".mpeg", ".m4v"]
-    )
-    return valid_http and valid_video
-
-
-def extract_video_from_links(links):
-    filtered_links = [{"url": link["url"], "alt": link.get("text", "")} for link in links if valid_video_link(link)]
-    return filtered_links
-
-
-text_extensions = set(
-    [
-        "pdf",
-        "epub",
-        "djvu",
-        "mobi",
-        "doc",
-        "docx",
-        "rtf",
-        "txt",
-        "odt",
-        "ppt",
-        "pptx",
-        "pages",
-        "keynote",
-        "wps",
-        "md",
-    ]
-)
-
-
-def valid_text_link(link):
-    if not link.get("url", "").startswith("http"):
-        return False
-    splits = link.get("url", "").split(".")
-    if len(splits) < 2:
-        return False
-    if splits[-1] not in text_extensions:
-        return False
-    return True
-
-
-def extract_text_from_links(links):
-    filtered_links = [{"url": link["url"], "alt": link.get("text", "")} for link in links if valid_text_link(link)]
-    return filtered_links
-
-
-def valid_audio_link(link):
-    valid_http = link.get("url", "").startswith("http")
-    valid_audio = any(link.get("url", "").endswith(ext) for ext in [".ogg", ".wav", ".mp3", ".flac", ".m4a"])
-    return valid_http and valid_audio
-
-
-def extract_audio_from_links(links):
-    """Extract image from links"""
-    filtered_links = [{"url": link["url"], "alt": link.get("text", "")} for link in links if valid_audio_link(link)]
-    return filtered_links
-
-
-def valid_image_link(link):
-    valid_path = link.get("path", "") == "IMG@/src"
-    valid_alt = len(link.get("alt", "")) > 0
-    valid_http = link.get("url", "").startswith("http")
-    return valid_path and valid_http and valid_alt
-
-
-def extract_image_from_links(links):
-    """Extract image from links"""
-    filtered_links = [{"url": link["url"], "alt": link["alt"]} for link in links if valid_image_link(link)]
-    return filtered_links
-
-
-def extract_documents_from_links(links, document_type):
-    """Extract documents from links ; this function returns a list of dict {"alt": ..., "url": ...}"""
-
-    if document_type == "image":
-        return extract_image_from_links(links)
-    elif document_type == "audio":
-        return extract_audio_from_links(links)
-    elif document_type == "text":
-        return extract_text_from_links(links)
-    elif document_type == "video":
-        return extract_video_from_links(links)
-    else:
-        raise ValueError(f"Unknown document type {document_type}")
-
-
-def extract_documents_from_wat(stream, document_type):
-    """Extract document from stream"""
-    all_links = []
-    try:
-        for record in ArchiveIterator(stream, record_types=WarcRecordType.metadata, parse_http=False):
-            try:
-                record_data = simdjson.load(record.reader)  # type: ignore
-            except:  # pylint: disable=bare-except
-                logger.info("A shard record failed")
-                continue
-            envelope = record_data["Envelope"]
-            payload = envelope["Payload-Metadata"]
-            if "HTTP-Response-Metadata" not in payload:
-                continue
-            http_resp = payload["HTTP-Response-Metadata"]
-            if "HTML-Metadata" not in http_resp:
-                continue
-            metadata = http_resp["HTML-Metadata"]
-            if "Links" not in metadata:
-                continue
-
-            links = metadata["Links"]
-
-            filtered_links = extract_documents_from_links(links, document_type)
-            for link in filtered_links:
-                link["uid"] = str(hashlib.md5((link["alt"] + link["url"]).encode()).hexdigest())
-            all_links.extend(filtered_links)
-    except Exception as e:  # pylint: disable=broad-except
-        logger.info(e)
-        logger.info("A shard failed to parse")
-        return []
-
-    return all_links
-
-
-def process_wat(path, document_type):
-    """Process a single wat file"""
-    begin_read = timer()
-    with fsspec.open(path, "rb") as f:
-        for i in range(10):
-            try:
-                tf = BytesIO(f.read())
-                break
-            except Exception as ex:  # pylint: disable=broad-except
-                if i == 9:
-                    logger.info("failed 10 times, skipping ", path)
-                    return
-                logger.info(ex)
-                logger.info(f"retrying reading {i}/10")
-                time.sleep(1)
-
-        for e in extract_documents_from_wat(tf, document_type):
-            yield (e["uid"], e["url"], e["alt"])
-    end_read = timer()
-    tot_read_time = end_read - begin_read
-    logger.info(f"Took {tot_read_time} to parse")
-
-
-def get_cc_wat_links(source_cc_protocol):
-    """Get cc wat links"""
-    if source_cc_protocol == "s3":
-        fs, p = fsspec.core.url_to_fs("s3://commoncrawl/crawl-data/")
-        links = ["s3://" + e for e in fs.glob(p + "/*/wat.paths.gz")]
-        return links
-    elif source_cc_protocol == "http":
-        fs, p = fsspec.core.url_to_fs("https://commoncrawl.org/the-data/get-started/")
-        a = fs.open(p).read()
-        l = a.splitlines()
-        l = [e.decode("utf8").replace("[WARC] ", "") for e in l]
-        l = [e for e in l if "<li>s3://commoncrawl/crawl-data/" in e]
-        l = [
-            e.split(" ")[0].replace("<li>s3://commoncrawl/", "https://data.commoncrawl.org/").replace("<wbr>", "")
-            for e in l
-        ]
-        l = [(e + "/wat.paths.gz").replace("//wat", "/wat") for e in l]
-        return l
-    else:
-        raise ValueError(f"Unknown protocol {source_cc_protocol}")
-
-
-def read_wat_index_file(wat_index):
-    with fsspec.open(wat_index, "rb", compression="gzip") as f:
-        wats = [a.decode("utf8").strip() for a in f.readlines()]
-    return wats
-
-
-def read_wat_index_files(shard_count, wat_count, source_cc_protocol):
-    """Read all wat index files"""
-    cc_wat_links = get_cc_wat_links(source_cc_protocol)
-    if shard_count is not None:
-        cc_wat_links = cc_wat_links[-shard_count:]  # pylint: disable=invalid-unary-operand-type
-    all_wats = []
-    with ThreadPool(16) as pool:
-        for wats in pool.imap_unordered(read_wat_index_file, cc_wat_links):
-            all_wats.extend(wats)
-    if wat_count is not None:
-        all_wats = random.choices(all_wats, k=wat_count)
-    else:
-        # shuffle to increase duplication over each part hence reduce size of each part after duplication
-        random.shuffle(all_wats)
-    return all_wats
 
 
 def deduplicate_repartition_count(df, output_path, wat_count, spark, shuffle=False):
@@ -227,25 +39,40 @@ def deduplicate_repartition_count(df, output_path, wat_count, spark, shuffle=Fal
     logger.info(f"Size: {df.count()}")
 
 
-def process_one_part(output_path, wat_index_files, build_spark, shuffle, document_type, source_cc_protocol):
+def process_one_part(output_path, cc_index_files, build_spark, shuffle, document_type, source_cc_protocol,ccfile):
     """Process one part"""
     spark = build_spark()
     sc = SparkContext.getOrCreate()
-    wat_count = len(wat_index_files)
-    wat_rdd = sc.parallelize(wat_index_files, wat_count)
+    ccfile_count = len(cc_index_files)
+    wat_rdd = sc.parallelize(cc_index_files, ccfile_count)
+
     if source_cc_protocol == "s3":
         prefix = "s3://commoncrawl/"
     elif source_cc_protocol == "http":
         prefix = "https://data.commoncrawl.org/"
 
-    def extract(x):
-        x = list(x)
-        yield from process_wat(prefix + x[0], document_type)
+    if ccfile == "warc":
+        def extract(x):
+            x = list(x)
+            yield from process_warc(prefix + x[0])
+
+    elif ccfile == "wat":
+        def extract(x):
+            x = list(x)
+            yield from process_wat(prefix + x[0], document_type)
+            
+    elif ccfile == "wet":
+        def extract(x):
+            x = list(x)
+            yield from process_wet(prefix + x[0])
+    else:
+        raise ValueError(f"Unknown ccfile: {ccfile}")
+
 
     output = wat_rdd.mapPartitions(extract)
     df = output.toDF(["uid", "url", "alt"])
 
-    deduplicate_repartition_count(df, output_path, wat_count, spark, shuffle)
+    deduplicate_repartition_count(df, output_path, ccfile_count, spark, shuffle)
 
 
 def get_last_successful_part(output_path):
@@ -258,7 +85,7 @@ def get_last_successful_part(output_path):
 
 
 def process_multi_part(
-    output_path, wat_index_files, build_spark, multipart, shuffle, resume, document_type, source_cc_protocol
+    output_path, cc_index_files, build_spark, multipart, shuffle, resume, document_type, source_cc_protocol,ccfile
 ):
     """Process multi part"""
     if resume:
@@ -266,16 +93,16 @@ def process_multi_part(
     else:
         start_part = 0
 
-    wat_count = len(wat_index_files)
-    wat_per_part = math.ceil(wat_count / multipart)
+    ccfile_count = len(cc_index_files)
+    ccfile_per_part = math.ceil(ccfile_count / multipart)
     part_paths = []
     for i in range(start_part, multipart):
-        start = i * wat_per_part
-        end = (i + 1) * wat_per_part
+        start = i * ccfile_per_part
+        end = (i + 1) * ccfile_per_part
         part_path = f"{output_path}/part_{i}"
         part_paths.append(part_path)
         logger.info(f"Processing part {i} from {start} to {end} into {part_path}")
-        process_one_part(part_path, wat_index_files[start:end], build_spark, False, document_type, source_cc_protocol)
+        process_one_part(part_path, cc_index_files[start:end], build_spark, False, document_type, source_cc_protocol,ccfile)
 
     spark = build_spark()
     logger.info("Merging parts")
@@ -287,7 +114,7 @@ def process_multi_part(
         else:
             df = df.union(spark.read.parquet(part_path))
 
-    deduplicate_repartition_count(df, output_path + "/merged", wat_count, spark, shuffle)
+    deduplicate_repartition_count(df, output_path + "/merged", ccfile_count, spark, shuffle)
 
 
 def get_date_str():
@@ -296,8 +123,8 @@ def get_date_str():
 
 def cc2dataset(
     output_path,
-    wat_index_count=1,
-    wat_count=100,
+    crawl_index_count=1,
+    files_count=100,
     master="local",
     num_cores=128,
     mem_gb=256,
@@ -307,6 +134,8 @@ def cc2dataset(
     spark_builder=None,
     document_type="image",
     source_cc_protocol="s3",
+    ccfile="wat",
+    crawl_index_list=None,
 ):
     """Convert common crawl to image caption set"""
 
@@ -332,20 +161,19 @@ def cc2dataset(
         return spark_builder()
 
     if resume is None:
-        wat_index_files = read_wat_index_files(wat_index_count, wat_count, source_cc_protocol)
-        # write wat index files to disk in output_path with fsspec
-        with fsspec.open(f"{output_path}/wat_index_files.txt", "w", encoding="utf8") as f:
-            f.write("\n".join(wat_index_files))
+        cc_index_files = read_index_files(crawl_index_count, files_count, source_cc_protocol, ccfile, crawl_index_list)
+        # write ccfile index files to disk in output_path with fsspec
+        with fsspec.open(f"{output_path}/crawl_index_files.txt", "w", encoding="utf8") as f:
+            f.write("\n".join(cc_index_files))
     else:
-        with fsspec.open(f"{output_path}/wat_index_files.txt", "r", encoding="utf8") as f:
-            wat_index_files = f.read().splitlines()
+        with fsspec.open(f"{output_path}/crawl_index_files.txt", "r", encoding="utf8") as f:
+            cc_index_files = f.read().splitlines()
 
     if multipart is None:
-        process_one_part(output_path, wat_index_files, build_spark, shuffle, document_type, source_cc_protocol)
+        process_one_part(output_path, cc_index_files, build_spark, shuffle, document_type, source_cc_protocol,ccfile)
     else:
         process_multi_part(
-            output_path, wat_index_files, build_spark, multipart, shuffle, resume, document_type, source_cc_protocol
-        )
+            output_path, cc_index_files, build_spark, multipart, shuffle, resume, document_type, source_cc_protocol, ccfile)
 
 
 def main():
